082225
Decent work on getting external clock to work, but had to sacrifice lights in order to do so. Apparently neopixels and midi clock input don't like to play nice? At least not with me doing more recalculations for the delta locks... it wasn't as much of a big deal when we were just locking down the existing steps, not populating new arrays constantly. Anyway, this is an interesting version with no visual feedback. converting from ws2812 to dotstar might help in the future, that or having a dedicated microcontroller for leds and another for the midi. options abound... this thing continues to throw curveballs